requirements.txt
Pillow==9.2.0
tk